Move CPU feature detection functions to processor.cpp
Use C++ instead of C since there will be a lot of string handling and indirect interaction
with LLVM in this file. The meta-programing capability is also pretty useful.
